refactored redux structure

+++ fruits_in_choco_frontend/src/components/Admin/Product/common/ProductTableActions.js
@@ -2,9 +2,9 @@ import React, {Component} from 'react';
 import PropTypes from 'prop-types'
 import {NavLink} from "react-router-dom";
 import {connect} from "react-redux";
-import {deleteProductById} from "../../../../redux/thunks/admin_thunks";
 import {compose} from "redux";
 import {withSnackbar} from "notistack";
+import {deleteProductById} from "../../../../redux/thunks/product_thunks";
 
 const propTypes = {
     row: PropTypes.object.isRequired,
@@ -12,7 +12,7 @@ const propTypes = {
 
 class ProductTableActions extends Component {
     handleDelete(e, id) {
-        if (window.confirm('Удлаить?')) {
+        if (window.confirm('Удлаить?') && !this.props.isProductDeleting) {
             this.props.deleteProductById(id);
         }
     }
